requirements.txt
matplotlib==3.3.1
numpy==1.19.1
scipy==1.5.2
pytest==6.0.1
networkx==2.5
pyyaml==5.3.1
requests==2.24.0
trimesh==3.8.3
triangle==20200424
pyglet==1.5.7
meshpy==2018.2.1
shapely==1.7.1
rtree==0.9.4